# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/presentations:
    get:
      operationId: DirectPresentationController_createPresentation
      summary: Request a presentation
      description: Generates a Veriable Presentation in Jwt or JSON-LD format for one
        of the credentials in this wallet
      parameters:
        - name: credentialId
          required: true
          in: query
          schema:
            type: string
        - name: asJwt
          required: true
          in: query
          schema:
            type: boolean
        - name: unwrap
          required: true
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/VerifiablePresentationJwtDto"
                  - $ref: "#/components/schemas/VerifiablePresentationJsonLdDto"
      tags: &a3
        - Presentation Direct
      security: &a4
        - oauth2:
            - wallet_view_presentations
  /api/presentations/validate:
    post:
      operationId: DirectPresentationController_validatePresentation
      summary: Validate presentation
      description: Validates a Jwt-based Verifiable Presentation according to a fixed
        set of requirements
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifiablePresentationJwtDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresentationValidationDto"
      tags: *a3
      security: *a4
  /api/iatp/holder/token:
    get:
      operationId: IatpHolderController_createSIToken
      summary: Request a SIOP token
      description: Generates a SIOP token for the provided audience to allow it to
        request presentations
      parameters:
        - name: audience
          required: true
          in: query
          schema:
            type: string
        - name: scope
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_token:
                    type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: &a5
        - Presentation IATP
      security:
        - oauth2:
            - wallet_view_presentations
  /api/iatp/holder/presentation:
    get:
      operationId: IatpHolderController_getPresentation
      summary: Retrieve presentation
      description: Request a presentation with a SIOP-token and a presentation definition
      parameters:
        - name: presentation_definition
          required: true
          in: query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresentationResponseDto"
          schema:
            type: string
        - name: Authorization
          required: true
          in: header
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresentationResponseDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a5
...
  /keys/{id}:
    get:
      operationId: KeysController_getCaChain
      summary: Retrieve Key CA chain
      description: Retrieves the CA chain for a given key. Only supported for keys
        with type `X509`
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags:
        - Keys
  /api/management/keys:
    get:
      operationId: KeysManagementController_getKeys
      summary: Retrieve keys
      description: Retrieves all keys registered for this wallet
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeyInfoDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: &a12
        - Management Keys
      security: &a13
        - oauth2:
            - wallet_manage_keys
    post:
      operationId: KeysManagementController_addKey
      summary: Add key
      description: Generates a new key based on the provided configuration
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyInfoDto"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "409":
          description: Resource already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a12
      security: *a13
  /api/management/keys/{keyId}:
    get:
      operationId: KeysManagementController_getKey
      summary: Retrieve key
      description: Retrieves key information of a specific key within this wallet
      parameters:
        - name: keyId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyInfoDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a12
      security: *a13
    delete:
      operationId: KeysManagementController_deleteKey
      summary: Delete key
      description: Deletes an existing key within this wallet
      parameters:
        - name: keyId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a12
      security: *a13
  /api/management/keys/{keyId}/default:
    put:
      operationId: KeysManagementController_setDefaultKey
      summary: Set default key
      description: Sets the provided key as default key within this wallet, will
        remove the default key flag for other keys in this wallet
      parameters:
        - name: keyId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a12
      security: *a13
...
info:
  title: TSG Wallet
  description: Description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          scopes:
            wallet_view_did: wallet_view_did
            wallet_manage_keys: wallet_manage_keys
            wallet_view_own_credentials: wallet_view_own_credentials
            wallet_view_all_credentials: wallet_view_all_credentials
            wallet_manage_own_credentials: wallet_manage_own_credentials
            wallet_manage_all_credentials: wallet_manage_all_credentials
            wallet_issue_credentials: wallet_issue_credentials
            wallet_view_presentations: wallet_view_presentations
  schemas:
...
    ErrorDto:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        code:
          type: number
        message:
          type: object
        error:
          type: string
      required:
        - name
        - status
        - code
    VerifiablePresentationJwtDto:
      type: object
      properties:
        vp:
          type: string
      required:
        - vp
    SignatureDto:
      type: object
      properties:
        type:
          type: string
        created:
          type: string
          format: date-time
        proofPurpose:
          type: string
        jws:
          type: string
        verificationMethod:
          type: string
      required:
        - type
        - created
        - proofPurpose
        - jws
        - verificationMethod
    VerifiableCredentialDto:
      type: object
      properties:
        "@context":
          type: array
          items:
            type: string
        type:
          type: array
          items:
            type: string
        id:
          type: string
        credentialSubject:
          oneOf:
            - $ref: "#/components/schemas/DefaultCredentialSubjectDto"
            - type: array
              items:
                $ref: "#/components/schemas/DefaultCredentialSubjectDto"
        issuer:
          type: string
        expirationDate:
          type: string
          format: date-time
        issuanceDate:
          type: string
          format: date-time
        evidence:
          type: object
        proof:
          $ref: "#/components/schemas/SignatureDto"
      required:
        - "@context"
        - type
        - credentialSubject
        - issuer
        - issuanceDate
        - proof
    VerifiablePresentationJsonLdDto:
      type: object
      properties:
        vp:
          $ref: "#/components/schemas/VerifiableCredentialDto"
      required:
        - vp
    PresentationValidationDto:
      type: object
      properties:
        valid:
          type: boolean
        validateJWTSignature:
          type: boolean
        validateJWTExpiryDate:
          type: boolean
        validateTrustAnchors:
          type: array
          items:
            type: boolean
        validateExpiryDate:
          type: array
          items:
            type: boolean
        validateCredentials:
          type: array
          items:
            type: boolean
        validateAudience:
          type: boolean
        vp:
          type: string
      required:
        - valid
        - validateJWTSignature
        - validateJWTExpiryDate
        - validateTrustAnchors
        - validateExpiryDate
        - validateCredentials
        - vp
...
    DescriptorMapDto:
      type: object
      properties:
        id:
          type: string
        format:
          type: string
        path:
          type: string
        path_nested:
          $ref: "#/components/schemas/DescriptorMapDto"
      required:
        - id
        - format
        - path
    PresentationSubmissionDto:
      type: object
      properties:
        id:
          type: string
        definition_id:
          type: string
        descriptor_map:
          type: array
          items:
            $ref: "#/components/schemas/DescriptorMapDto"
      required:
        - id
        - definition_id
        - descriptor_map
    PresentationResponseDto:
      type: object
      properties:
        vp_token:
          type: string
        presentation_submission:
          $ref: "#/components/schemas/PresentationSubmissionDto"
      required:
        - vp_token
        - presentation_submission
...
    DefaultCredentialSubjectDto:
      type: object
      properties:
        id:
          type: string
      required:
        - id
...
    KeyType:
      type: string
      enum:
        - EdDSA
        - ES384
        - X509
    JsonWebKeyDto:
      type: object
      properties:
        alg:
          type: string
        crv:
          type: string
        e:
          type: string
        ext:
          type: boolean
        key_ops:
          type: array
          items:
            type: string
        kid:
          type: string
        kty:
          type: string
        n:
          type: string
        use:
          type: string
        x:
          type: string
        y:
          type: string
      required:
        - kty
    KeyInfoDto:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/KeyType"
        default:
          type: boolean
        publicKey:
          $ref: "#/components/schemas/JsonWebKeyDto"
        created:
          format: date-time
          type: string
        modified:
          format: date-time
          type: string
      required:
        - id
        - type
        - default
        - publicKey
        - created
        - modified
    KeyConfigDto:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/KeyType"
        id:
          type: string
        default:
          type: boolean
        existingKey:
          type: string
        existingCertificate:
          type: string
      required:
        - type
        - id
        - default
...